Fix compiler issues due to SofwareSerial
When SerialFirmata was extracted from StandardFirmataPlus to a
separate class it broke the ability to compile Firmata with
versions of the Arduino IDE older than 1.6.6. This is because
prior to 1.6.6 Arduino did not allow libraries to include other
libraries unless the other library was included in the sketch.
In this particular case that was not possible because
SerialFirmata is compiled before the sketch is compiled.

There were a couple of potential work arounds:
1. Only include SoftwareSerial when compiling in Arduino 1.6.6 or
   higher.
2. Split SerialFirmata into SoftwareSerialFirmata and
   HardwareSerialFirmata and then include SoftwareSerial in the
   sketch if SoftwareSerialFirmata is included. However this
   requires the user to modify the sketch.

I opted for option 1. If users really need to use software serial
with SerialFirmata in Arduino 1.6.5 or older, I can reconsider
option 2 at a later time.