Add support for Arduino MKR WiFi 1010
The board uses the U-Blox WiFi NINA chip which has an ESP32 inside running
NINA firmware. The SAMD21 is connected to the U-Blox chip via SPI. The
WiFININA library is used to communicate between the U-Blox chip and the
SAMD21. Other boards use the same U-Blox chip such as the MKR Vidor 4000
and the Uno WiFi Rev2. Additional changes will be required to support
those boards.